+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support Drupal 10 and farmOS v3
+
 ## [2.17.0](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/milestone/37) 2023-11-02
 
 ### Added